+++ assets/src/edit-story/app/media/local/initIndexDb.js
